requirements.txt
discord.py>=2.0.0
gspread>=5.0.0
google-auth>=2.0.0
python-dotenv>=0.19.0